Extract JFixture as separate extension
