+++ Heroku/backend/database/src/test/kotlin/net/twisterrob/cinema/database/model/UserTest.kt
@@ -3,15 +3,16 @@ package net.twisterrob.cinema.database.model
 import com.flextrade.jfixture.JFixture
 import com.shazam.shazamcrest.MatcherAssert.assertThat
 import com.shazam.shazamcrest.matcher.Matchers.sameBeanAs
-import net.twisterrob.test.applyCustomisation
 import net.twisterrob.test.build
 import net.twisterrob.test.buildList
 import org.hamcrest.Matchers.sameInstance
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.extension.ExtendWith
 
+@ExtendWith(ModelFixtureExtension::class)
 class UserTest {
 
-	private val fixture: JFixture = JFixture().applyCustomisation { add(validDBData()) }
+	private lateinit var fixture: JFixture
 
 	@Test fun `copy creates a full copy`() {
 		val fixtFilm: Film = fixture.build()
