+++ Heroku/backend/database/src/test/kotlin/net/twisterrob/cinema/database/model/ViewIntgTest.kt
@@ -22,7 +22,7 @@ import org.neo4j.ogm.session.Session
 import org.neo4j.ogm.session.loadAll
 import java.time.temporal.ChronoUnit
 
-@ExtendWith(ModelIntgTestExtension::class)
+@ExtendWith(ModelIntgTestExtension::class, ModelFixtureExtension::class)
 @TagIntegration
 class ViewIntgTest {
