+++ Heroku/backend/endpoint/src/test/kotlin/net/twisterrob/cinema/cineworld/backend/endpoint/film/data/QuickbookFilmRepositoryTest.kt
@@ -6,29 +6,26 @@ import com.nhaarman.mockitokotlin2.verifyNoMoreInteractions
 import com.nhaarman.mockitokotlin2.whenever
 import net.twisterrob.cinema.cineworld.backend.endpoint.view.data.View
 import net.twisterrob.cinema.cineworld.quickbook.QuickbookService
-import net.twisterrob.cinema.database.model.validDBData
+import net.twisterrob.cinema.database.model.ModelFixtureExtension
 import net.twisterrob.cinema.database.services.FilmService
-import net.twisterrob.test.applyCustomisation
 import net.twisterrob.test.build
-import net.twisterrob.test.offsetDateTimeRealistic
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.extension.ExtendWith
 import org.mockito.Mock
 import org.mockito.MockitoAnnotations
 import net.twisterrob.cinema.cineworld.backend.endpoint.film.data.Film as FrontendFilm
 import net.twisterrob.cinema.database.model.Film as DBFilm
 
+@ExtendWith(ModelFixtureExtension::class)
 class QuickbookFilmRepositoryTest {
 
 	@Mock lateinit var mockService: FilmService
 	@Mock lateinit var mockMapper: FilmMapper
 	@Mock lateinit var mockQuickbook: QuickbookService
 
-	private val fixture = JFixture().applyCustomisation {
-		add(validDBData())
-		add(offsetDateTimeRealistic())
-	}
+	private lateinit var fixture: JFixture
 	private lateinit var sut: FilmRepository
 
 	@BeforeEach fun setUp() {
