+++ Heroku/backend/sync/src/test/kotlin/net/twisterrob/cinema/cineworld/sync/CinemaSyncTest.kt
@@ -5,7 +5,7 @@ import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.whenever
 import net.twisterrob.cinema.cineworld.sync.syndication.Feed
-import net.twisterrob.cinema.database.model.validDBData
+import net.twisterrob.cinema.database.model.ModelFixtureExtension
 import net.twisterrob.cinema.database.services.CinemaService
 import net.twisterrob.test.applyCustomisation
 import net.twisterrob.test.build
@@ -15,16 +15,18 @@ import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.containsInAnyOrder
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.extension.ExtendWith
 import java.time.OffsetDateTime
 import net.twisterrob.cinema.database.model.Cinema as DBCinema
 
+@ExtendWith(ModelFixtureExtension::class)
 class CinemaSyncTest {
 
 	private val calculator: CinemaSyncCalculator = mock()
 	private val dbService: CinemaService = mock()
 	private val now by lazy { OffsetDateTime.now() }
 
-	private val fixture = JFixture()
+	private lateinit var fixture: JFixture
 	private lateinit var sut: CinemaSync
 
 	@BeforeEach fun setUp() {
@@ -34,7 +36,6 @@ class CinemaSyncTest {
 	@Test fun `all changed cinemas are synced properly`() {
 		fixture.applyCustomisation {
 			add(syncResults<DBCinema>())
-			add(validDBData())
 		}
 
 		val fixtFeed: Feed = fixture.build()
