+++ Heroku/backend/sync/src/test/kotlin/net/twisterrob/cinema/cineworld/sync/CinemaSyncUpdaterTest.kt
@@ -9,14 +9,14 @@ import com.flextrade.jfixture.JFixture
 import com.shazam.shazamcrest.matcher.Matchers.sameBeanAs
 import dagger.Component
 import net.twisterrob.cinema.cineworld.sync.syndication.Feed
-import net.twisterrob.cinema.database.model.validDBData
-import net.twisterrob.test.applyCustomisation
+import net.twisterrob.cinema.database.model.ModelFixtureExtension
 import net.twisterrob.test.assertAll
 import net.twisterrob.test.build
 import net.twisterrob.test.that
 import org.hamcrest.Matchers.equalTo
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.extension.ExtendWith
 import net.twisterrob.cinema.cineworld.sync.syndication.Feed.Cinema as FeedCinema
 import net.twisterrob.cinema.database.model.Cinema as DBCinema
 
@@ -24,11 +24,10 @@ import net.twisterrob.cinema.database.model.Cinema as DBCinema
  * @see SyncAppModule.copyCinemaProperties sut
  * @see DBCinema.copyPropertiesFrom delegate
  */
+@ExtendWith(ModelFixtureExtension::class)
 class CinemaSyncUpdaterTest {
 
-	private val fixture = JFixture().applyCustomisation {
-		add(validDBData())
-	}
+	private lateinit var fixture: JFixture
 	private lateinit var sut: Updater<DBCinema, FeedCinema>
 
 	@BeforeEach fun setUp() {
