+++ Heroku/backend/sync/src/test/kotlin/net/twisterrob/cinema/cineworld/sync/FilmSyncTest.kt
@@ -5,7 +5,7 @@ import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.whenever
 import net.twisterrob.cinema.cineworld.sync.syndication.Feed
-import net.twisterrob.cinema.database.model.validDBData
+import net.twisterrob.cinema.database.model.ModelFixtureExtension
 import net.twisterrob.cinema.database.services.FilmService
 import net.twisterrob.test.applyCustomisation
 import net.twisterrob.test.build
@@ -15,16 +15,18 @@ import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.containsInAnyOrder
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.extension.ExtendWith
 import java.time.OffsetDateTime
 import net.twisterrob.cinema.database.model.Film as DBFilm
 
+@ExtendWith(ModelFixtureExtension::class)
 class FilmSyncTest {
 
 	private val calculator: FilmSyncCalculator = mock()
 	private val dbService: FilmService = mock()
 	private val now by lazy { OffsetDateTime.now() }
 
-	private val fixture = JFixture()
+	private lateinit var fixture: JFixture
 	private lateinit var sut: FilmSync
 
 	@BeforeEach fun setUp() {
@@ -34,7 +36,6 @@ class FilmSyncTest {
 	@Test fun `all changed cinemas are synced properly`() {
 		fixture.applyCustomisation {
 			add(syncResults<DBFilm>())
-			add(validDBData())
 		}
 
 		val fixtFeed: Feed = fixture.build()
