+++ Heroku/backend/sync/src/test/kotlin/net/twisterrob/cinema/cineworld/sync/PerformanceSyncUpdaterTest.kt
@@ -10,14 +10,14 @@ import com.shazam.shazamcrest.matcher.Matchers.sameBeanAs
 import dagger.Component
 import net.twisterrob.cinema.cineworld.sync.syndication.Feed
 import net.twisterrob.cinema.cineworld.sync.syndication.Feed.Companion.DEFAULT_TIMEZONE
-import net.twisterrob.cinema.database.model.validDBData
-import net.twisterrob.test.applyCustomisation
+import net.twisterrob.cinema.database.model.ModelFixtureExtension
 import net.twisterrob.test.assertAll
 import net.twisterrob.test.build
 import net.twisterrob.test.that
 import org.hamcrest.Matchers.equalTo
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.extension.ExtendWith
 import net.twisterrob.cinema.cineworld.sync.syndication.Feed.Performance as FeedPerformance
 import net.twisterrob.cinema.database.model.Performance as DBPerformance
 
@@ -25,11 +25,10 @@ import net.twisterrob.cinema.database.model.Performance as DBPerformance
  * @see SyncAppModule.copyPerformanceProperties sut
  * @see DBPerformance.copyPropertiesFrom delegate
  */
+@ExtendWith(ModelFixtureExtension::class)
 class PerformanceSyncUpdaterTest {
 
-	private val fixture = JFixture().applyCustomisation {
-		add(validDBData())
-	}
+	private lateinit var fixture: JFixture
 	private lateinit var sut: Updater<DBPerformance, FeedPerformance>
 
 	@BeforeEach fun setUp() {
